fix: Fix message event exception
Message events may not have an origin, and they may not have an originalEvent field.  This happens in particular when loading Karma and the Chromecast SDK on a Chromecast.  An event arrives from the platform with neither of these properties.

This fixes the error by checking for originalEvent before accessing a property of it.